+++ inc/system/requirement/extensiongroup.class.php
@@ -37,46 +37,53 @@ if (!defined('GLPI_ROOT')) {
 }
 
 /**
- * @since 9.5.0
+ * @since 10.0.0
  */
-class Extension extends AbstractRequirement {
+class ExtensionGroup extends AbstractRequirement {
 
    /**
-    * Required extension name.
+    * Required extensions names.
     *
-    * @var string
+    * @var string[]
     */
-   protected $name;
+   protected $extensions;
 
    /**
-    * @param string      $name         Required extension name.
+    * @param string      $name         Extension group title.
+    * @param string[]    $extensions   Required extensions names.
     * @param bool        $optional     Indicate if extension is optional.
     * @param string|null $description  Describe usage of the extension.
     */
-   public function __construct(string $name, bool $optional = false, ?string $description = null) {
-      $this->title = sprintf(__('%s extension'), $name);
-      $this->name = $name;
+   public function __construct(string $title, array $extensions, bool $optional = false, ?string $description = null) {
+      $this->title = $title;
+      $this->extensions = $extensions;
       $this->optional = $optional;
       $this->description = $description;
    }
 
    protected function check() {
-      $this->validated = extension_loaded($this->name);
-      $this->buildValidationMessage();
-   }
+      $loaded_extensions  = [];
+      $missing_extensions = [];
 
-   /**
-    * Defines the validation message based on self properties.
-    *
-    * @return void
-    */
-   protected function buildValidationMessage() {
-      if ($this->validated) {
-         $this->validation_messages[] = sprintf(__('%s extension is installed.'), $this->name);
-      } else if ($this->optional) {
-         $this->validation_messages[] = sprintf(__('%s extension is not present.'), $this->name);
-      } else {
-         $this->validation_messages[] = sprintf(__('%s extension is missing.'), $this->name);
+      foreach ($this->extensions as $extension) {
+         if (extension_loaded($extension)) {
+            $loaded_extensions[] = $extension;
+         } else {
+            $missing_extensions[] = $extension;
+         }
+      }
+
+      $this->validated = count($missing_extensions) === 0;
+
+      if (count($loaded_extensions) > 0) {
+         $this->validation_messages[] = sprintf(__('Following extensions are installed: %s.'), implode(', ', $loaded_extensions));
+      }
+      if (count($missing_extensions) > 0) {
+         if ($this->optional) {
+            $this->validation_messages[] = sprintf(__('Following extensions are not present: %s.'), implode(', ', $missing_extensions));
+         } else {
+            $this->validation_messages[] = sprintf(__('Following extensions are missing: %s.'), implode(', ', $missing_extensions));
+         }
       }
    }
 }